third_party/blink/web_tests/platform/generic/http/tests/forms/form-submission-cancelable-expected.txt

Test that form submit within onsubmit event handlers are not delayed and sends the form data when invoked

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.target.contentWindow.location.search is "?query=AAA"
PASS event.target.contentWindow.location.search is "?query1=BBB"
PASS successfullyParsed is true

TEST COMPLETE

 
 
 
